+++ monitor/hmp-cmds.c
@@ -1348,7 +1348,7 @@ void hmp_balloon(Monitor *mon, const QDict *qdict)
 void hmp_nballoon(Monitor *mon, const QDict *qdict)
 {
     int64_t value = qdict_get_int(qdict, "value");
-    int64_t node = qdict_get_int(qdict, "node");
+    int32_t node = qdict_get_int(qdict, "node");
     Error *err = NULL;
 
     qmp_nballoon(value, node, &err);
